+++ tests/examples.rs
@@ -7,3 +7,15 @@ mod quickstart_example;
 #[allow(missing_docs)]
 #[path = "../examples/custom_matcher.rs"]
 mod custom_matcher_example;
+
+#[allow(missing_docs)]
+#[path = "../examples/parameterized.rs"]
+mod parameterized_example;
+
+#[allow(missing_docs)]
+#[path = "../examples/setup_inheritance.rs"]
+mod setup_inheritance_example;
+
+#[allow(missing_docs)]
+#[path = "../examples/teardown.rs"]
+mod teardown_example;